+++ src/main/java/team/gdsc/code2cv/feature/resume/domain/career/CareerListConverter.java
@@ -1,4 +1,4 @@
-package team.gdsc.code2cv.feature.resume.domain;
+package team.gdsc.code2cv.feature.resume.domain.career;
 
 import java.util.List;
